Take advantage of the fact that all texture units have a separate texture matrix
git-svn-id: https://svn.il.fontys.nl/usersvn/fishtank/svn/trunk@330 bb1b31de-baa9-489e-af1c-03370f2903b8

+++ fishtank/Ground.h
@@ -31,7 +31,7 @@ class Ground
 		boost::gil::gray32f_image_t heightmap;
 		Texture texture;
 		std::vector<Texture> caustics;
-		TriangleArray<unsigned int, float, float, float, 2> triangles;
+		TriangleArray<unsigned int, float, int, float, 2> triangles;
 };
 
 #endif // __INCLUDED_GROUND_H__
